Reverted motor axis move
Part of the redesign work had moved the lead screws closer together to
better fit the NEMA 23 mounts on the top which required some redesign of
the gantry. This commit attaches NEMA 23 in a way which does not require
the lead screws to be moved so the existing instrument is much easier to
bring up to date. The only downside is that the NEMA 23 mounting plate
is not directly bolted to the vertical beams (this was an improvement on
the current design which does not do this and it will still be possible
to bolt the new DC motor and stepper motor plates to the verticals just
not the NEMA 23). If this proves to be a problem then it would be
trivial to make one if thicker material was used.